test_requirements.txt
# Core dependencies for PaddleOCR testing
paddlepaddle==3.0.0b1
paddleocr[all]>=3.3.0

# Additional testing utilities
pytest>=7.0.0
pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
numpy>=1.21.0

# For document parsing tests
pdf2image>=1.16.0

